requirements.txt
# Core Framework & API
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
typing_extensions==4.15.0
anyio==4.11.0
sniffio==1.3.1

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
PyJWT==2.10.1
cryptography==46.0.3
ecdsa==0.19.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# Environment & Config
python-dotenv==1.1.1

# Date/Time utilities
python-dateutil==2.9.0.post0
pytz==2025.2

# Payment processing
razorpay==2.0.0

# PDF Generation
reportlab==4.4.4
pillow==12.0.0

# HTTP requests
httpx==0.28.1
requests==2.32.5
httpcore==1.0.9
h11==0.16.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0

# Email validation
email-validator==2.3.0

# Utility dependencies (required by main packages)
six==1.17.0
cffi==2.0.0
pycparser==2.23
